feat: collation job code
