+++ src/components/projects/workflow/workflowEditor/customWorkflow/jobOperate.vue
@@ -29,14 +29,9 @@ export default {
       curTaskIndex: -1
     }
   },
-  props: {
-    jobInfo: {
-      type: Object,
-      default: () => ({})
-    }
-  },
   computed: {
     taskList () {
+      // 由两部分组成 一部分前端定义 + 后端插件列表（并且插件需要自己造结构）
       return this.jobTypeList.concat(this.plugins)
     }
   },
